+++ Bus Scheduler/package.json
@@ -13,10 +13,10 @@
   "dependencies": {
     "axios": "^1.7.2",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.1"
   },
   "devDependencies": {
-    "normalize.css": "^8.0.1",
     "@testing-library/jest-dom": "^6.4.5",
     "@testing-library/react": "^15.0.7",
     "@types/node": "^20.12.12",
@@ -29,6 +29,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.6",
     "jsdom": "^24.0.0",
+    "normalize.css": "^8.0.1",
     "sass": "^1.77.2",
     "typescript": "^5.2.2",
     "vite": "^5.2.0",
