+++ Bus Scheduler/src/App.tsx
@@ -1,5 +1,7 @@
-function App() {
+import { FC } from 'react';
+
+const App: FC = () => {
     return 'Site would be up shortly';
-}
+};
 
 export default App;
