+++ Bus Scheduler/src/components/atoms/image/index.tsx
