+++ Bus Scheduler/src/components/layouts/home/components/style.scss
@@ -1,0 +1,4 @@
+.content-container
+{
+    height: 100%;
+}
