+++ Bus Scheduler/src/components/layouts/tripList/components/style.scss
@@ -1,0 +1,4 @@
+.content-container
+{
+    height: 100%;
+}
