+++ Bus Scheduler/src/components/molecules/inputWithIcon/style.scss
