+++ Bus Scheduler/src/components/molecules/logo/style.scss
