+++ Bus Scheduler/src/components/organisms/footer/index.tsx
@@ -1,0 +1,9 @@
+import '@/components/organisms/footer/style.scss';
+import '@/components/organisms/footer';
+import { COPY_RIGHT } from './constants';
+
+const Footer = () => {
+    return <div className="footer-container"> {COPY_RIGHT}</div>;
+};
+
+export default Footer;
