+++ Bus Scheduler/src/components/organisms/footer/style.scss
