+++ Bus Scheduler/src/components/organisms/tripSearch/store/storeState.ts
