+++ Bus Scheduler/src/components/organisms/tripSearch/store/types.ts
