+++ Bus Scheduler/src/components/templates/style.scss
