+++ Bus Scheduler/src/main.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import Home from '@/features/home';
+import AppRouter from '@/routes/AppRouter';
 import '@/index.scss';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
         <div className="app-container">
-            <Home></Home>
+            <AppRouter></AppRouter>
         </div>
     </React.StrictMode>
 );
