--- Bus Scheduler/src/shared/components/atoms/link/index.tsx
@@ -1,16 +1,0 @@
-import '@/shared/components/atoms/link/style.scss';
-
-type LinkType = {
-    text?: string;
-    url: string;
-    children?: React.ReactNode | undefined;
-};
-const Link = ({ text, url, children }: LinkType) => {
-    return (
-        <a href={url}>
-            {text} {children && children}
-        </a>
-    );
-};
-
-export default Link;
