--- Bus Scheduler/src/shared/components/organisms/footer/index.tsx
@@ -1,9 +1,0 @@
-import '@/shared/components/organisms/footer/style.scss';
-import '@/shared/components/organisms/footer';
-import { COPY_RIGHT } from './constants';
-
-const Footer = () => {
-    return <div className="footer-container"> {COPY_RIGHT}</div>;
-};
-
-export default Footer;
